+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## 1.1.0 - Unreleased
+- Added `Genemo.crossover.edgeRecombination` function.
 - Added type checking to `elitism`, `evaluatePopulation`, `generateInitialPopulation`, `stopCondition` and all selection functions.
 
 ## 1.0.1 - 2019-07-27